+++ p5-server/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Libraries can be specified as comment directives
 - Performance improvements: cache script analysis
 - Directory index ignores additional patterns
+- Switch back to parsing the HTML in order to inject the script tag; fall back
+  with a warning
 
 ## [0.8.5] - 2021-10-28